Separate AppBatContent
